requirements.txt
Flask
requests
PyMuPDF
python-docx
pytesseract
Pillow
langdetect
werkzeug
